requirements.txt
# Common
python-dotenv
pydantic==2.*
requests

# Data Ingestion
google-cloud-bigquery

# Analytics
pandas
sqlalchemy
psycopg2-binary==2.9.9
functions-framework==3.* # Correct package name
flask

# API
fastapi
uvicorn[standard]

# Slack Bot
slack_sdk

# Dashboard
streamlit
altair

# Testing
pytest
numpy
mypy
pandas-stubs
types-requests
types-psycopg2
ruff 